requirements.txt
# Core API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
websockets>=12.0

# Database
sqlalchemy>=2.0.25

# File parsing
PyPDF2>=3.0.1
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2
Pillow>=10.2.0
python-magic>=0.4.27
beautifulsoup4>=4.12.0
markdown>=3.5.0

# LLM & Embeddings (cloud providers - optional)
anthropic>=0.18.1
openai>=1.12.0
tiktoken>=0.6.0
numpy>=1.24.0

# File watching
watchdog>=4.0.0

# HTTP client
httpx>=0.26.0

# CLI
click>=8.1.7
rich>=13.7.0

# Utilities
aiofiles>=23.2.1
python-dotenv>=1.0.1
tqdm>=4.66.1
